% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{absnorm}
\alias{absnorm}
\title{Absolute normalization of a vector}
\usage{
absnorm(vec, max.norm = FALSE)
}
\arguments{
\item{vec}{Vector to normalize}

\item{max.norm}{Logical, whether to use maximum normalization}
}
\value{
Normalized vector
}
\description{
Absolute normalization of a vector
}
