% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{cluster_ensemble_workflow}
\alias{cluster_ensemble_workflow}
\title{Complete clustering ensemble workflow}
\usage{
cluster_ensemble_workflow(
  train_data,
  test_data,
  n_clusters = 5,
  outcome_col = "y",
  merged_trees = 500,
  cluster_trees = 100,
  cluster_ind = TRUE,
  n_cores = 2,
  seed = NULL
)
}
\arguments{
\item{train_data}{Either a single dataframe or a list of pre-clustered dataframes}

\item{test_data}{Either a single dataframe or a list of test dataframes}

\item{n_clusters}{Number of clusters to create if train_data is a single dataframe}

\item{outcome_col}{Name of outcome column}

\item{merged_trees}{Number of trees for merged model}

\item{cluster_trees}{Number of trees for individual cluster models}

\item{cluster_ind}{Boolean representing if we should perform k-means clustering on the merged train_data or keep the original structure}

\item{n_cores}{Number of cores for parallel processing (default = 2)}

\item{seed}{Random seed for reproducibility}
}
\value{
List containing fitted models, predictions, and performance metrics
}
\description{
Complete clustering ensemble workflow
}
