% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{create_clusters}
\alias{create_clusters}
\title{Create clusters from training data}
\usage{
create_clusters(train_data, n_clusters, outcome_col = "y")
}
\arguments{
\item{train_data}{Training dataset (tibble/data.frame)}

\item{n_clusters}{Number of clusters to create}

\item{outcome_col}{Name of outcome column}
}
\value{
List containing cluster assignments and clustered data
}
\description{
Create clusters from training data
}
