% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{crosscluster_predict}
\alias{crosscluster_predict}
\title{Make predictions using fitted CCWF models}
\usage{
crosscluster_predict(ccwf_fit, new_data)
}
\arguments{
\item{ccwf_fit}{Output from crosscluster_fit function}

\item{new_data}{New data for predictions}
}
\value{
Tibble with predictions from all methods
}
\description{
Make predictions using fitted CCWF models
}
