% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{plot_performance_comparison}
\alias{plot_performance_comparison}
\title{Plot performance comparison across methods}
\usage{
plot_performance_comparison(workflow_results)
}
\arguments{
\item{workflow_results}{Output from cluster_ensemble_workflow}
}
\value{
ggplot object
}
\description{
Plot performance comparison across methods
}
