% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_data_monte}
\alias{sim_data_monte}
\title{Simulate non-gaussian clusters using the 'monte' function}
\usage{
sim_data_monte(nclusters, ncoef, ntest)
}
\arguments{
\item{nclusters}{Number of clusters to simulate for the training dataset}

\item{ncoef}{Number of coefficients}

\item{ntest}{Number of test datasets to simulate}
}
\value{
Returns a list with two components:
\item{clusters_list}{A list of training data tibbles, one per cluster. Each tibble contains:
\itemize{
\item y: The simulated outcome variable generated using linear, quadratic, and interaction terms
\item V1...Vn: The simulated predictor variables, where n equals ncoef, generated using monte carlo methods
}
}
\item{test_list}{A list of test data tibbles. Each tibble has the same structure as the training data}
}
\description{
Simulate non-gaussian clusters using the 'monte' function
}
