% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{absnorm}
\alias{absnorm}
\title{Absolute normalization of a vector}
\usage{
absnorm(vec, max.norm = FALSE)
}
\arguments{
\item{vec}{Numeric vector to normalize}

\item{max.norm}{Logical, whether to use maximum normalization (default: FALSE)}
}
\value{
Returns a numeric vector of the same length as the input, where each element is:
1. Multiplied by its original sign
2. Normalized by dividing by the sum of absolute differences from either:
- The minimum value (if max.norm = FALSE)
- The maximum value (if max.norm = TRUE)
}
\description{
Absolute normalization of a vector
}
\examples{
absnorm(c(-2, 1, 3))
absnorm(c(-2, 1, 3), max.norm = TRUE)
}
