% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{crosscluster_fit}
\alias{crosscluster_fit}
\title{Fit cross-cluster weighted forests method}
\usage{
crosscluster_fit(
  train_data,
  outcome_col = "y",
  model_fit = rf_fit,
  model_predict = rf_predict,
  merged_trees = 500,
  cluster_trees = 100,
  cluster_ind = TRUE,
  n_cores = 2
)
}
\arguments{
\item{train_data}{Either a single dataframe or a list of pre-clustered dataframes}

\item{outcome_col}{Name of outcome column}

\item{model_fit}{Function to fit models (defaults to rf_fit)}

\item{model_predict}{Function to make predictions (defaults to rf_predict)}

\item{merged_trees}{Number of trees for merged model}

\item{cluster_trees}{Number of trees for individual cluster models}

\item{cluster_ind}{Boolean representing if we should perform k-means clustering on the merged train_data or keep the original structure}

\item{n_cores}{Number of cores for parallel processing (default = 2)}
}
\value{
List of fitted models and weights
}
\description{
Fit cross-cluster weighted forests method
}
