% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{norm_vec}
\alias{norm_vec}
\title{Normalize a vector}
\usage{
norm_vec(x)
}
\arguments{
\item{x}{Numeric vector to normalize}
}
\value{
Normalized vector with Euclidean (L2) norm
}
\description{
Normalize a vector
}
\examples{
norm_vec(c(3, 4)) # Should return 5
}
