% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{rf_predict}
\alias{rf_predict}
\title{Make Predictions with a Fitted Random Forest Model}
\usage{
rf_predict(model, newdata)
}
\arguments{
\item{model}{A fitted model object from rf_fit()}

\item{newdata}{A data frame or tibble containing new data for prediction}
}
\value{
Returns a numeric vector containing the predicted values for the new data
}
\description{
Make Predictions with a Fitted Random Forest Model
}
